fixed linters issues fixed code review issues
