+++ aquality_selenium_core/exceptions/__init__.py
@@ -1,0 +1,1 @@
+"""Package for сustom exceptions."""
