+++ aquality_selenium_core/exceptions/illegal_argument_exception.py
@@ -1,2 +1,7 @@
+"""Module for illegal argument exception."""
+
+
 class IllegalArgumentException(Exception):
+    """Class for illegal argument exception."""
+
     pass
