--- aquality_selenium_core/utilities/abstract_action_retrier.py
@@ -1,17 +1,0 @@
-from abc import ABC
-from abc import abstractmethod
-from typing import Any
-from typing import Callable
-from typing import List
-from typing import Type
-from typing import Union
-
-
-class AbstractActionRetrier(ABC):
-    @abstractmethod
-    def do_with_retry(
-        self,
-        function_and_args: List[Union[Callable, Any]],
-        handled_exceptions: List[Type[Exception]],
-    ) -> Any:
-        pass
