+++ aquality_selenium_core/utilities/resource_file.py
@@ -1,5 +1,11 @@
+"""Module defines work with getting data from files."""
+
+
 class ResourceFile:
+    """Class, which defines getting data from file."""
+
     def __init__(self, path_to_resource: str):
+        """Initialize resource file by provided path."""
         self.__file_canonical_path = path_to_resource
         self.__file_content = self.__get_resource_file_content()
 
@@ -9,4 +15,9 @@ class ResourceFile:
 
     @property
     def file_content(self) -> str:
+        """
+        Return file content.
+
+        :return: File content.
+        """
         return self.__file_content
