+++ setup.cfg
@@ -17,7 +17,6 @@ classifiers =
     Operating System :: POSIX
     Programming Language :: Python :: 3
     Programming Language :: Python :: 3 :: Only
-    Programming Language :: Python :: 3.5
     Programming Language :: Python :: 3.6
     Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
@@ -37,7 +36,7 @@ packages =
 install_requires =
     jsonpath-ng==1.5.1
     selenium==3.141.0
-python_requires = >=3.5
+python_requires = >=3.6
 package_dir =
     =aquality_selenium_core
 setup_requires =
