+++ tests/utilities_tests/settings_file_tests/custom_settings_file_tests.py
@@ -3,15 +3,16 @@ from typing import Dict
 from typing import List
 
 from hamcrest import assert_that
+from hamcrest import none
 
-from aquality_selenium_core.utilities.abstract_settings_file import AbstractSettingsFile
+from aquality_selenium_core.utilities.settings_file import AbstractSettingsFile
 
 
 class TestCustomSettingsFile:
     def test_should_be_possible_to_override_settings_file(self):
         custom_implementation: AbstractSettingsFile = CustomSettingsFile()
         value = custom_implementation.get_value("timeouts.timeoutPollingInterval")
-        assert_that(value is None, "Value should be got from CustomSettingsFile")
+        assert_that(value, none(), "Value should be got from CustomSettingsFile")
 
 
 class CustomSettingsFile(AbstractSettingsFile):
